added styles

+++ comments/views.py
@@ -2,12 +2,13 @@ from django.shortcuts import get_object_or_404, redirect
 from django.views.generic import DetailView, View
 from .models import Comment
 from .forms import CommentForm
-from posts.models import Post
+from posts.models import Post, PostLike
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.urls import reverse_lazy
 # from notifications.utils import process_mentions
 from comments.models import CommentLike
 from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
 
 
 class PostDetailWithCommentsView(LoginRequiredMixin, DetailView):
@@ -21,7 +22,10 @@ class PostDetailWithCommentsView(LoginRequiredMixin, DetailView):
         context = super().get_context_data(**kwargs)
 
         comments = self.object.comments.all().order_by("-created_at")
-
+        post = self.object
+        post.post_liked_by_current_user = False
+        if self.request.user.is_authenticated:
+            post.post_liked_by_current_user = PostLike.objects.filter(post=post, user=self.request.user).exists()
         # додаємо лайкнуто чи ні
         for c in comments:
             c.liked_by_current_user = CommentLike.objects.filter(
@@ -99,8 +103,21 @@ class CommentDeleteView(UserPassesTestMixin, View):
     
 @login_required
 def toggle_like(request, comment_id):
-    comment = get_object_or_404(Comment, pk=comment_id)
-    like, created = CommentLike.objects.get_or_create(comment=comment, user=request.user)
-    if not created:
-        like.delete()  # якщо лайк вже був — видаляємо (toggle)
-    return redirect("post:post-detail", comment.post.uuid)
+    if request.method == "POST" and request.headers.get("X-Requested-With") == "XMLHttpRequest":
+        comment = get_object_or_404(Comment, pk=comment_id)
+        like, created = CommentLike.objects.get_or_create(comment=comment, user=request.user)
+        if not created:
+            like.delete()  # toggle
+
+        data = {
+            "liked": comment.likes.count() if like.pk else False,
+            "likes_count": comment.likes.count(),
+            "user_liked": comment.likes.filter(user=request.user).exists()
+        }
+        return JsonResponse({
+            "liked": like.pk is not None,  # True якщо лайк створений
+            "likes_count": comment.likes.count()
+        })
+
+    # якщо не AJAX або не POST
+    return JsonResponse({"error": "Invalid request"}, status=400)
